requirements.txt
gradio>=3.50.2
python-dotenv>=1.0.0
oci>=2.110.0
aiofiles>=23.2.1
aiohttp>=3.8.5
pillow>=10.0.0
asyncio>=3.4.3